[Dialogue] Add function to update Town UI
