+++ Assets/Scripts/Systems/NPCDataSystem-Josie/NPCUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 59866aed15e72024db75d86cf4a919ab
