+++ Assets/Scripts/Systems/NPCDataSystem-Josie/TownUI.cs
@@ -1,16 +1,21 @@
 using UnityEngine;
-
+using TMPro;
 public class TownUI : MonoBehaviour
 {
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
+    [Header("Town UI references")]
+    [SerializeField] NPCUI npcUI;
+    [SerializeField] Trading.TradingUI tradingUI;
+    NPCData npcData = null;
+
+
+    // THIS NEEDS TO BE CALLED WHEN A PLAYER CLICKS A TOWN
+    public void SetTownUI(NPCData newData)
     {
+        npcData = newData;
+        npcUI.DisplayNPCUI(npcData);
+        tradingUI.SetTown(npcData.townStock);
         
     }
 
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
+    
 }
